+++ neurolabi/gui/flyem/zflyembodymergeproject.h
@@ -66,7 +66,7 @@ public:
   }
 
   //Obsolete functions
-  int getSelectedBodyId() const;
+  uint64_t getSelectedBodyId() const;
   void addSelected(uint64_t label);
   void removeSelected(uint64_t label);
